+++ hubdle-project-gradle-plugin/test/kotlin/com/javiersc/hubdle/project/config/documentation/changelog/ChangelogMergedTest.kt
@@ -1,9 +1,9 @@
 package com.javiersc.hubdle.project.config.documentation.changelog
 
-import com.javiersc.hubdle.project._utils.resourceFile
 import com.javiersc.hubdle.project.extensions.config.documentation.changelog._internal.Changelog
 import com.javiersc.hubdle.project.extensions.config.documentation.changelog._internal.fromString
 import com.javiersc.hubdle.project.extensions.config.documentation.changelog._internal.merged
+import com.javiersc.hubdle.project.fixtures.resourceFile
 import io.kotest.matchers.shouldBe
 import kotlin.test.Test
 
@@ -25,7 +25,7 @@ private fun assert(number: Int) {
 private const val BASE_PATH = "config/documentation/changelog"
 
 private fun getChangelog(number: Int): String =
-    resourceFile("$BASE_PATH/changelog/merge-$number/CHANGELOG.md").readText()
+    resourceFile("$BASE_PATH/merge-$number/CHANGELOG.md").readText()
 
 private fun expectChangelog(number: Int): String =
-    resourceFile("$BASE_PATH/changelog/merge-$number/CHANGELOG_EXPECT.md").readText()
+    resourceFile("$BASE_PATH/merge-$number/CHANGELOG_EXPECT.md").readText()